feat: adding permissions check before publishing to disconnected registry
Fixes #196
