+++ pkg/cli/mirror/catalog_images_test.go
@@ -20,6 +20,7 @@ import (
 func Test_BuildCatalogLayer(t *testing.T) {
 
 	server := httptest.NewServer(registry.New())
+	t.Cleanup(server.Close)
 	u, err := url.Parse(server.URL)
 	if err != nil {
 		t.Error(err)
